Separated filters to their own file and created context to provide active filter to item filtering logic
