+++ src/Home.tsx
@@ -1,10 +1,14 @@
-import ItemsToShowAndFilters from "./components/ItemsToShowAndFilters/ItemsToShowAndFilters.tsx";
+import ItemsToDisplay from "./components/ItemsToShowAndFilters/ItemsToDisplay.tsx";
+import { Filters } from "./components/Filters/Filters.tsx";
 
 const Home = () => {
   return (
     <>
       <div>
-        <ItemsToShowAndFilters />
+        <div className={"filter"}>
+          <Filters />
+        </div>
+        <ItemsToDisplay />
       </div>
     </>
   );
