+++ src/components/Filters/Filters.scss
@@ -1,0 +1,9 @@
+@import "../../scss/base";
+
+.filter {
+  margin-top: 60px;
+  margin-bottom: 60px;
+  margin-left: 60px;
+  display: flex;
+  gap: 24px;
+}
